fix(admin): security와 관련된 파일을 global로 이동
