+++ backend/src/main/java/me/performancereservation/global/security/admin/handler/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package me.performancereservation.domain.admin.handler;
+package me.performancereservation.global.security.admin.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
@@ -9,7 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import me.performancereservation.global.exception.ErrorCode;
 import me.performancereservation.global.exception.ErrorResponse;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
@@ -17,16 +17,16 @@ import java.io.IOException;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     private final ObjectMapper objectMapper;
 
-    //어드민의 인증정보(아이디, 비밀번호)가 틀린 경우의 예외를 처리하는 핸들러
+    //로그인을 하지 않고 admin 기능에 접근한 경우의 예외를 처리하는 핸들러
     @Override
-    public void onAuthenticationFailure(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        AuthenticationException exception) throws IOException, ServletException {
-        ErrorCode errorCode = ErrorCode.ADMIN_NOT_FOUND;
+    public void commence(HttpServletRequest request,
+                         HttpServletResponse response,
+                         AuthenticationException authException) throws IOException, ServletException {
+        ErrorCode errorCode = ErrorCode.ADMIN_AUTHENTICATION_REQUIRED;
         ErrorResponse errorResponse = ErrorResponse.from(errorCode);
 
         //response에 예외를 셋팅해 반환해 줍니다.
